refactor : refatorando arquivos de componentes
